Refactor MessageChunker
