+++ src/bot/commands/SoundsCommand.ts
@@ -4,17 +4,15 @@ import Config from '@config/Config';
 import localize from '@util/i18n/localize';
 import { getSounds } from '@util/SoundUtil';
 import Command from './base/Command';
-import MessageChunker from './helpers/MessageChunker';
+import { chunkedMessages } from './helpers/chunkedMessages';
 
 export default class SoundsCommand implements Command {
   public readonly TRIGGERS = ['sounds'];
 
   private readonly config: Config;
-  private readonly chunker: MessageChunker;
 
-  constructor(config: Config, chunker: MessageChunker) {
+  constructor(config: Config) {
     this.config = config;
-    this.chunker = chunker;
   }
 
   public run(message: Message, params: string[]) {
@@ -26,8 +24,6 @@ export default class SoundsCommand implements Command {
     }
 
     const page = parseInt(params[0]);
-
-    this.chunker.chunkedMessages(sounds, page)
-      .forEach(chunk => message.author.send(chunk));
+    chunkedMessages(sounds, page).forEach(chunk => message.author.send(chunk));
   }
 }
