+++ src/bot/commands/TagsCommand.ts
@@ -3,25 +3,17 @@ import { Message } from 'discord.js';
 import * as soundsDb from '@util/db/Sounds';
 import { getSounds } from '@util/SoundUtil';
 import Command from './base/Command';
-import MessageChunker from './helpers/MessageChunker';
+import { chunkedMessages } from './helpers/chunkedMessages';
 
 export default class TagsCommand implements Command {
   public readonly TRIGGERS = ['tags'];
-
-  private readonly chunker: MessageChunker;
-
-  constructor(chunker: MessageChunker) {
-    this.chunker = chunker;
-  }
 
   public run(message: Message, params: string[]) {
     const sounds = getSounds();
     const soundsWithTags = this.formattedMessage(sounds);
 
     const page = parseInt(params[0]);
-
-    this.chunker.chunkedMessages(soundsWithTags, page)
-      .forEach(chunk => message.author.send(chunk));
+    chunkedMessages(soundsWithTags, page).forEach(chunk => message.author.send(chunk));
   }
 
   private formattedMessage(sounds: string[]) {
